# Individual-Project/Backend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TourMate Backend API",
    "description": "OpenAPI 3.0 specification for the TourMate backend routes and controllers.",
    "version": "1.0.0",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "TourMate Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/tours": {
      "get": {
        "tags": ["Tours"],
        "summary": "Get tour requests for providers or tourists depending on role",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/SuccessTourRequestList" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["Tours"],
        "summary": "Create a tour request",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TourRequestCreate" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/SuccessTourRequest" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/bookings": {
      "post": {
        "tags": ["Bookings"],
        "summary": "Create a booking",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BookingCreate" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/SuccessBooking" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/payments": {
      "post": {
        "tags": ["Payments"],
        "summary": "Create a booking payment checkout",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PaymentCreate" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/SuccessPaymentCheckout" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/feedback": {
      "post": {
        "tags": ["Feedback"],
        "summary": "Create feedback",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "example": { "booking": "66b6c6f1f8a8a2b9f8b11111", "rating": 5, "comment": "Great service" }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/SuccessMessage" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/services": {
      "get": {
        "tags": ["Services"],
        "summary": "Get all services",
        "responses": {
          "200": { "$ref": "#/components/responses/SuccessServiceList" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["Services"],
        "summary": "Create a service",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ServiceCreate" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/SuccessService" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/sos": {
      "post": {
        "tags": ["SOS"],
        "summary": "Create an SOS alert",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SOSAlertCreate" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/SuccessMessage" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "get": {
        "tags": ["SOS"],
        "summary": "Get SOS alerts",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/SuccessSOSAlertList" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/admin/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "List all users",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": { "type": "object", "additionalProperties": true }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "SuccessMessage": {
        "description": "Success response with message",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/MessageResponse" }
          }
        }
      },
...
      "SuccessTourRequest": {
        "description": "Tour request response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/TourRequestResponse" }
          }
        }
      },
      "SuccessTourRequestList": {
        "description": "Tour request list response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/TourRequestListResponse" }
          }
        }
      },
      "SuccessBooking": {
        "description": "Booking response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/BookingResponse" }
          }
        }
      },
...
      "SuccessPaymentCheckout": {
        "description": "Booking payment checkout response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/PaymentCheckoutResponse" }
          }
        }
      },
...
      "SuccessService": {
        "description": "Service response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ServiceResponse" }
          }
        }
      },
      "SuccessServiceList": {
        "description": "Service list response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ServiceListResponse" }
          }
        }
      },
      "SuccessSOSAlertList": {
        "description": "SOS alert list response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/SOSAlertListResponse" }
          }
        }
      },
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "ServerError": {
        "description": "Server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string", "example": "Validation failed" },
          "error": { "type": "object", "nullable": true, "additionalProperties": true },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "data": { "type": "object", "additionalProperties": true }
        }
      },
...
      "TourRequestResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string" },
          "data": {
            "type": "object",
            "properties": {
              "tourRequest": { "$ref": "#/components/schemas/TourRequest" }
            }
          }
        }
      },
      "TourRequestListResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "tourRequests": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/TourRequest" }
              },
              "count": { "type": "integer", "example": 1 }
            }
          }
        }
      },
      "BookingResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "booking": { "$ref": "#/components/schemas/Booking" }
            }
          }
        }
      },
...
      "PaymentCheckoutResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string", "example": "Payment created" },
          "data": {
            "type": "object",
            "properties": {
              "payment": { "$ref": "#/components/schemas/Payment" }
            }
          }
        }
      },
...
      "ServiceResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "service": { "$ref": "#/components/schemas/Service" }
            }
          }
        }
      },
      "ServiceListResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "services": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/Service" }
              },
              "count": { "type": "integer", "example": 1 }
            }
          }
        }
      },
      "SOSAlertListResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "TourRequestCreate": {
        "type": "object",
        "required": ["title", "destinations", "startDate", "endDate", "travelers", "budget"],
        "properties": {
          "title": { "type": "string" },
          "destinations": { "type": "array", "items": { "type": "string" } },
          "districts": { "type": "array", "items": { "type": "string" } },
          "locationPlan": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "district": { "type": "string" },
                "locations": { "type": "array", "items": { "type": "string" } }
              }
            }
          },
          "startDate": { "type": "string", "format": "date" },
          "endDate": { "type": "string", "format": "date" },
          "travelers": { "type": "integer", "minimum": 1 },
          "visitorBreakdown": {
            "type": "object",
            "properties": {
              "male": { "type": "integer", "minimum": 0 },
              "female": { "type": "integer", "minimum": 0 },
              "kids": { "type": "integer", "minimum": 0 }
            }
          },
          "budget": {
            "type": "object",
            "required": ["min", "max"],
            "properties": {
              "min": { "type": "number" },
              "max": { "type": "number" },
              "currency": { "type": "string", "example": "USD" }
            }
          },
          "preferences": {
            "type": "object",
            "properties": {
              "accommodation": { "type": "string" },
              "transportation": { "type": "string" },
              "activities": { "type": "array", "items": { "type": "string" } },
              "servicesNeeded": { "type": "array", "items": { "type": "string" } },
              "dietary": { "type": "string" },
              "specialRequirements": { "type": "string" }
            }
          }
        }
      },
...
      "BookingCreate": {
        "type": "object",
        "required": ["service", "bookingDate", "numberOfPeople"],
        "properties": {
          "service": { "type": "string" },
          "tourRequest": { "type": "string", "format": "objectId" },
          "bookingDate": {
            "type": "object",
            "required": ["startDate", "endDate"],
            "properties": {
              "startDate": { "type": "string", "format": "date" },
              "endDate": { "type": "string", "format": "date" }
            }
          },
          "numberOfPeople": { "type": "integer", "minimum": 1 },
          "specialRequests": { "type": "string" }
        }
      },
...
      "PaymentCreate": {
        "type": "object",
        "required": ["bookingId"],
        "properties": {
          "bookingId": { "type": "string", "format": "objectId" },
          "paymentMethod": { "type": "string", "enum": ["stripe", "card", "bank-transfer", "manual"] },
          "paymentDetails": { "type": "object", "additionalProperties": true }
        }
      },
...
      "ServiceCreate": {
        "type": "object",
        "required": ["name", "type", "description", "pricing"],
        "properties": {
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["villa", "hotel", "vehicle", "guide", "restaurant", "safari", "photographer", "equipment", "other"] },
          "description": { "type": "string" },
          "location": {
            "type": "object",
            "properties": {
              "address": { "type": "string" },
              "city": { "type": "string" },
              "province": { "type": "string" },
              "coordinates": {
                "type": "object",
                "properties": {
                  "latitude": { "type": "number" },
                  "longitude": { "type": "number" }
                }
              }
            }
          },
          "pricing": {
            "type": "object",
            "required": ["amount"],
            "properties": {
              "amount": { "type": "number" },
              "currency": { "type": "string", "example": "USD" },
              "unit": { "type": "string", "enum": ["night", "day", "hour", "trip", "person", "item"] }
            }
          }
        }
      },
      "SOSAlertCreate": {
        "type": "object",
        "required": ["location", "emergencyType", "description", "contactNumber"],
        "properties": {
          "location": {
            "type": "object",
            "required": ["latitude", "longitude"],
            "properties": {
              "latitude": { "type": "number" },
              "longitude": { "type": "number" },
              "address": { "type": "string" }
            }
          },
          "emergencyType": { "type": "string", "enum": ["medical", "accident", "theft", "lost", "other"] },
          "description": { "type": "string" },
          "contactNumber": { "type": "string" }
        }
      },
...
      "Booking": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "format": "objectId" },
          "tourist": { "$ref": "#/components/schemas/User" },
          "service": { "type": "string" },
          "serviceSnapshot": { "type": "object", "additionalProperties": true },
          "provider": { "$ref": "#/components/schemas/User" },
          "tourRequest": { "type": "string", "format": "objectId" },
          "bookingDate": {
            "type": "object",
            "properties": {
              "startDate": { "type": "string", "format": "date-time" },
              "endDate": { "type": "string", "format": "date-time" }
            }
          },
          "numberOfPeople": { "type": "integer" },
          "pricing": { "type": "object", "additionalProperties": true },
          "status": { "type": "string" },
          "paymentStatus": { "type": "string" },
          "paymentId": { "type": "string", "format": "objectId" },
          "specialRequests": { "type": "string" },
          "cancellationReason": { "type": "string" },
          "cancelledBy": { "type": "string", "format": "objectId" },
          "cancelledAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "TourRequest": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "format": "objectId" },
          "tourist": { "$ref": "#/components/schemas/User" },
          "title": { "type": "string" },
          "destinations": { "type": "array", "items": { "type": "string" } },
          "districts": { "type": "array", "items": { "type": "string" } },
          "locationPlan": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "startDate": { "type": "string", "format": "date-time" },
          "endDate": { "type": "string", "format": "date-time" },
          "travelers": { "type": "integer" },
          "visitorBreakdown": { "type": "object", "additionalProperties": true },
          "budget": { "type": "object", "additionalProperties": true },
          "preferences": { "type": "object", "additionalProperties": true },
          "status": { "type": "string", "enum": ["open", "awaiting-payment", "in-progress", "completed", "cancelled"] },
          "bids": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "provider": { "$ref": "#/components/schemas/User" },
                "message": { "type": "string" },
                "proposedPrice": { "type": "number" },
                "itinerary": { "type": "string" },
                "submittedAt": { "type": "string", "format": "date-time" },
                "status": { "type": "string", "enum": ["pending", "accepted", "rejected"] }
              }
            }
          },
          "selectedBid": { "type": "string", "format": "objectId" },
          "advancePayment": {
            "type": "object",
            "properties": {
              "status": { "type": "string", "enum": ["not-required", "pending", "paid", "failed"] },
              "amount": { "type": "number" },
              "currency": { "type": "string" },
              "paymentId": { "type": "string", "format": "objectId" }
            }
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Service": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "format": "objectId" },
          "provider": { "$ref": "#/components/schemas/User" },
          "name": { "type": "string" },
          "type": { "type": "string" },
          "description": { "type": "string" },
          "location": { "type": "object", "additionalProperties": true },
          "pricing": { "type": "object", "additionalProperties": true },
          "capacity": { "type": "integer" },
          "images": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "amenities": { "type": "array", "items": { "type": "string" } },
          "availability": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "rating": { "type": "number" },
          "reviewCount": { "type": "integer" },
          "isActive": { "type": "boolean" },
          "isVerified": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Payment": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "format": "objectId" },
          "booking": { "type": "string", "format": "objectId" },
          "tourRequest": { "type": "string", "format": "objectId" },
          "tourist": { "$ref": "#/components/schemas/User" },
          "provider": { "$ref": "#/components/schemas/User" },
          "amount": { "type": "number" },
          "currency": { "type": "string" },
          "gateway": { "type": "string" },
          "gatewayOrderId": { "type": "string" },
          "gatewayStatusCode": { "type": "string" },
          "gatewaySignature": { "type": "string" },
          "paymentMethod": { "type": "string", "enum": ["stripe", "card", "bank-transfer", "manual"] },
          "purpose": { "type": "string", "enum": ["booking", "tour-request-advance"] },
          "transactionId": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "processing", "completed", "failed", "refunded"] },
          "paymentDetails": { "type": "object", "additionalProperties": true },
          "refund": { "type": "object", "additionalProperties": true },
          "metadata": { "type": "object", "nullable": true, "additionalProperties": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}